backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
Pillow==10.0.1
google-generativeai==0.2.0
python-dotenv==1.0.0
